partially implement Disk related backend code.

+++ cloudstorage/src/main/java/uk/ac/herts/jw17aca/cloudstorage/mapper/UserMapper.java
@@ -4,10 +4,10 @@ import uk.ac.herts.jw17aca.cloudstorage.pojo.User;
 
 public interface UserMapper {
 	
-	  public User selectById(int id);
+	  public User selectById(long id);
 	  public User selectByEmail(String email);
 	  public void add(User user);
-	  public void delete(int id);
+	  public void delete(long id);
 	  public void update(User user);
 	  
 }
